Move OperationAuditTrail to web package
